From Adam Sulmicki <adam[AT]cfar.umd.edu>: 	Add more type values for EAP. 	Fix off-by-one bug when displaying Code of EAP message.
Get rid of an unnecessary "volatile".

Give the code and type fields value_string arrays, and use them when
putting the code and type into the protocol tree.

Base the decision of whether to put the type field into the tree on the
request code, not on the length of the packet.

Display the Type-Data field, under that name, under the EAP tree, as
it's part of the EAP PDU.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@4779 f5534014-38df-0310-8fa8-9805f1628bb7