Add support for desegmentation of X11 messages.
Try to check whether the first request we see is an initial connection
request, and dissect it as such if so.

Get rid of the global "next_offset" variable.

Check for buffer overflows when building the summary item for a list
of keysyms.

Display BASE_DEC items in decimal, not hex.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@5173 f5534014-38df-0310-8fa8-9805f1628bb7