Fix the test in "dissect_smb_datetime()" for "no date/time specified" values.
Note that in a Negotiate Protocol response, the primary domain won't be
present if the negotiated dialect isn't "DOS LANMAN 2.1" or "LANMAN2.1".

At least for Info Standard replies for Transaction2 Find First2
requests, if the request had the "return resume keys" flag set, the
reply will have a resume key at the beginning of each entry.  We assume
that to be the case for Info Query EA Size and Info QUery EAs From List;
it does *not* appear to be the case for Find File Directory Info, Find
File Full Directory Info, or Find File Both Directory Info (they don't
have it even if the flag is set, at least in the captures I've seen).

The length of the name string in Find First2 entries doesn't include the
terminating '\0'; count that as well.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@5259 f5534014-38df-0310-8fa8-9805f1628bb7