Add #defines for the bits in the flag fields, and a macro to test whether a connection-oriented PDU is fragmented or not.
Clean up the handling of fragmented connection-oriented PDUs (the code
to handle fragmented PDUs can assume that it is not the case that both
PFC_FIRST_FRAG and PFC_LAST_FRAG are set, as that's an unfragmented
PDU).  Put an entry into the protocol tree for the fragment data in
fragmented PDUs.

For fragmented connectionless PDUs, don't hand the payload of any
fragment other than the first fragment to the subdissector.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@5687 f5534014-38df-0310-8fa8-9805f1628bb7