SIGINT is, at least as I read some Microsoft documentation, not supported in Win32 applications; use the native Win32 mechanism for catching ^C (and other events that would terminate a program running in a console window).
That mechanism (and the signal mechanism in the MSVC++ C run-time) cause
the handler to be run in a separate thread, so it can't just do a
longjmp.  Fortunately, WinPcap's packet-dispatch loop, unlike the
libpcap loop on some UNIX platforms, can be interrupted by ^C, so we
don't have to do the longjmp there - we can just set "ld.go" to FALSE to
terminate the capture loop.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@5776 f5534014-38df-0310-8fa8-9805f1628bb7