From Devin Heitmueller: dissect NTLMSSP authentication messages, and handle the flags field in NTLMSSP messages as a 32-bit field.
Make "get_unicode_or_ascii_string()" take a "Unicode or not" flag rather
than a "packet_info *" as an argument, make it not static, and move it
to "packet-smb-common.c", so that it can be used by the SMB dissector
and the NTLMSSP dissector.  Also get rid of some _U_'s that are applied
to arguments that are, in fact, used.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@5976 f5534014-38df-0310-8fa8-9805f1628bb7