Fixed a bitwise vs logical AND bug when checking the ntlmssp flags for the unicode bit.
Also, it seems that the strings in the address list of a
NTLMSSP_CHALLENGE message are always in unicode, regardless of the
negotiated string type.  I have a capture of win98 doing NTLM over
HTTP where the domain name is in ASCII but the address list is
unicode.

There is still a bug in the dissection of the NTLMSSP_AUTH message
where the flags value does not specify unicode but the
domain/user/host name is unicode.  Perhaps the flags value for this
message aren't NTLMSSP flags?

Guy/Richard/jmayer, if you have any captures that show different
behaviour can you send them my way?


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@6329 f5534014-38df-0310-8fa8-9805f1628bb7