Don't include the header file to get the SNMP version unless we're building with an SNMP library.
If we have Net-SNMP, include <net-snmp/version.h>, not
<ucd-snmp/version.h>.

Don't include any of the SNMP headers unless HAVE_SOME_SNMP is defined.

Include <net-snmp/config_api.h> if we have Net-SNMP, to declare
"read_premib_configs()" and "read_configs()".

Supply the include directories for Net-SNMP in the Makefile.nmake for
GTK 1.2 and GTK 2.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@6493 f5534014-38df-0310-8fa8-9805f1628bb7