There's no need to include "prefs.h" unless we're actually exporting the big table of functions - we don't use the preferences code, we just export routines from it.
There is, however, a need to include "reassemble.h" if we're doing so,
as some of the functions we export take a "fragment_data *" as an
argument, and "fragment_data" is defined in "reassemble.h".

The members of "patable" should be set to point to functions, and
"p_prefs_register_obsolete_preference" isn't a function -
"prefs_register_obsolete_preference" is.

Get rid of extra white space at the ends of lines.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@6648 f5534014-38df-0310-8fa8-9805f1628bb7