Get rid of global variables - pass them as arguments.
Rename the "osi_decode_" routines "ositp_decode_", to indicate that
they're for OSI COTP and CLTP.

Fetch the dst_ref field in each dissector, rather than globally, as not
all PDUs have it.  In particular, DT PDUs don't have it in classes 0 and
1, so don't put it into the summary line for class 0 and 1 COTP.

Register the CLNP dissector for use by the X.25 dissector.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@6849 f5534014-38df-0310-8fa8-9805f1628bb7