It appears that the first two bytes of "xxz" are, in fact, the actual length of the packet, and the second two bytes are the captured length of the packet.  The old "length" value appears to be the captured length of the packet as well; perhaps it's to be interpreted as the number of bytes of data following the packet header (just in case there's padding, for example).
Treat "ATM/", as an encapsulation string, as RFC 1483 ATM.  (It may
actually be raw ATM, but the only capture I've seen had, in the parts I
saw, only RFC 1483 traffic LLC/SNAP traffic.)

There are 8 bytes in front of the LLC/SNAP header in ATM captures; skip
them, for now.  (Perhaps they're a pseudo-header, giving VPI/VCI
information and stuff such as that?  Or perhaps that's in the record
header?)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@6871 f5534014-38df-0310-8fa8-9805f1628bb7