Set the length of the L2TP tvbuff based on the length in the L2TP header, if present.
Don't explicitly base the length of the subset we hand to the PPP
dissector on the length in the header, as it might not be present;
instead, just use -1 for the lengths, as we set the length of our tvbuff
if there's a length in the L2TP header.

Set the length of the L2TP top-level protocol item for data packets, so
it covers the L2TP header but not the packet data.

Don't put a padding item into the tree if there's no padding.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@6922 f5534014-38df-0310-8fa8-9805f1628bb7