FileSize appears to be big-endian in DOSFileEntryStruct and FileInstance structures; make it so (which makes it big-endian on other structures as well).
The same appears to be true of the UpdateID.

There doesn't appear to be a reserved word between the
SearchSequenceWord and DirectoryID fields of a DirectoryInstance or a
FileInstance.

The dates and times in a DirectoryInstance or a FileInstance appear to
be big-endian; note that we need to add the ability to make those
instances big-endian but leave others little-endian.

Note that there's a sequence of NameSpaceNames and a sequence of
DataStreamNames in the reply to a Get Name Space Operation request.

A Login Object request has a 2-byte object type followed by a
ClientName, just a UserName.  The same applies to a Get Object
Connection List request.

Note that the DirectoryInstance or FileInstance in a File Search
Continue reply appears to be garbage if the reply doesn't succeed with
"OK".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7077 f5534014-38df-0310-8fa8-9805f1628bb7