Make the dissector static - it's not called from outside packet-bofl.c
Let the tvbuff mechanism check the length of the packet - don't check it
ourselves.

Put each field into the protocol tree and the Info column separately, so
that we at least get a partial dissection - I've seen some packets that
look like breath-of-life packets but that have only the PDU field.

Show the PDU field with all 8 hex digits, and show the sequence number
as an unsigned quantity, in the Info column.

Show the padding size based on the actual length of the frame, not on
the amount of the frame that was captured.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7238 f5534014-38df-0310-8fa8-9805f1628bb7