When reassembling, test the correct M bit based on whether it's mod-8 or mod-128.
Don't reassemble if we don't have all the payload (e.g., if we have a
short frame).

Clean up indentation.

Remove the "_x25" at the end of the preference for reassembling X.25
packets - it already has "x25." at the beginning.

Add a bunch of #defines for various bits to be tested, and use them.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7279 f5534014-38df-0310-8fa8-9805f1628bb7