Make various variables containing unsigned quantities unsigned.
Check whether a device identifier in a vital product data page (not a
vital data product page, although I guess most SCSI devices could be
considered data products) is bigger than the page before displaying it.

Don't bother to check whether the full identifier is in the packet - if
it's not, an exception will be thrown and an error indication put into
the protocol tree, which is the correct thing to do.  Do, however, check
whether the identifier length is 0 before putting a protocol tree entry
in.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7290 f5534014-38df-0310-8fa8-9805f1628bb7