When registering a string preference, if the value of the preference is NULL, convert it to a copy of a null string, otherwise replace it with a copy of the string, so that we know that the variable for the preference always points to a string that can be freed.
That also obviates the need to worry about a null-pointer value for a
preference variable when checking to see whether a preference has changed.

When checking for a string preference not being set, check for an empty
string, not a null pointer - the above code turns null pointers into
pointers to empty strings, *and* the GUI code does (and always did!) the
same.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7342 f5534014-38df-0310-8fa8-9805f1628bb7