Make the individual flag bits in the Connection Control field of the SPX header be filterable fields.
Don't hand retransmitted SPX frames to subdissectors - just show the
payload as a retransmission of the original frame.

Instead of handing a retransmission indicator to SPX subdissectors, hand
them a structure containing the datastream type (under the assumption
that it's data for the protocol running atop SPX, and that the dissector
for that protocol might use it) and the state of the end-of-message bit
(under the assumption that it's data for the protocol running atop SPX).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7433 f5534014-38df-0310-8fa8-9805f1628bb7