Use symbolic names for the various write mode bits.
To test whether a single bit is set, just do "if (mode&bit)", not
"if ((mode&bit)==bit)".

In the places where read and write data is processed, have both a
comment indicating that it's file data and that you can transport DCERPC
over SMB just with reads and writes, to indicate why we may call the
DCERPC-over-a-pipe dissector.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7450 f5534014-38df-0310-8fa8-9805f1628bb7