Make the dialog box "transient for" the main window, as we do with other pop-up dialogs.
If thee's already one such dialog, and the user asks for another one,
reactivate the old one (so it gets un-minimized if necessary), don't
just raise it.

Put the "Create Stat" button in a box, so it doesn't widen to the full
width of the window, and add a "Cancel" button to dismiss the dialog
box.

Arrange that <ESC> dismisses the dialog box, and that typing <Enter> in
the filter dialog box activates the dialog box.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7553 f5534014-38df-0310-8fa8-9805f1628bb7