Heuristic dissectors are not allowed to return FALSE after they've done something to the protocol tree or the columns, as that leaves crud in the protocol tree that could mess up whatever stuff subsequent dissection code does with the packet.
Get rid of all column-setting code before the initial sanity checking
code, and have that code just return FALSE rather than putting
"Malformed FIX Packet" indications (if the dissector returns FALSE, it's
saying the packet *isn't* a FIX packet, not that it is one but that it's
malformed).  After we've set the columns and created the protocol tree,
return TRUE if we find a problem (we should put an error indication
there in that case).


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7857 f5534014-38df-0310-8fa8-9805f1628bb7