Break out the code for the ahndling of the Service Response Time (formerly known as RoundTripTime) table and change the table into a scrollable CList.
As a first conversion to use the helper routines, convert DCERPC SRT statistics to use the new interface.
This prevents some interfaces (SAMR/LSA) that contains a huge number of procedures from creating a huge table that does not fir on the screen.

Later changes to the helpers may be to make the different columns sortable
or to hide those procedures that has not been seen in the capture.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7903 f5534014-38df-0310-8fa8-9805f1628bb7