remove one more redundant decode of the extensions present bit.
change the decode of sequence and extensions  to assume the lower bound for the number of extensions is 1 and thus 1 have to be added to the encoded value.

dont know if this is right or not, the satndard x691 does not mention anything
about the lb being 1 and the value being semiconstrained but a note at 18.8
does mention that the number of extensions can not be 0.

i think there is a difference between saying a value can not be zero and
between saying the lower bound is 1.  but hey it is a telco standard.

the change might be right or it might be wrong.
i think it is wrong or else the standard is wrong.
it at least dissects the very few captures i have properly.
telco guys, either give feedback or live with the dissector being potentially
wrong.
its that easy.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7990 f5534014-38df-0310-8fa8-9805f1628bb7