Replace "hf_scsi_log_pc" with separate "hf_scsi_logsel_pc" and "hf_scsi_logsns_pc" fields, with the latter using "scsi_logsns_pc_val[]" as its value_string table (the names are different).
Rename "scsi_modesense_medtype_val[]" to
"scsi_modesense_medtype_sbc_val[]", as it's specific to block devices;
add in additional information from a SCSI-3 Block Commands draft, and
have "dissect_scsi_modeselect6()" and "dissect_scsi_modeselect10()" use
it if the device is a block device.

Get rid of "scsi_inq_evpd_val[]", as it's not used and is redundant (we
already have "scsi_evpd_pagecode_val[]".  Add the Implemented operating
definition page to the latter table.

Fix up a "val_to_str()" call to include a format string in the "if
unknown" format, so the value is shown.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@7995 f5534014-38df-0310-8fa8-9805f1628bb7