Add a routine "dissect_ipv6_options()" that works like "dissect_ip_tcp_options()" but for options that are like IPv6 options (i.e., the length byte has a value that doesn't include the option code or length byte).
Add an "ip_opts.h" header to declare it, and move the declaration of
stuff used by it and "dissect_ip_tcp_options()", and the declaration of
"dissect_ip_tcp_options()", to that header.

Use "dissect_ipv6_options()" for Mobile IPv6 options.

Get rid of the unused "mip6_opt_types[]" array in "packet-mip6.h".


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8015 f5534014-38df-0310-8fa8-9805f1628bb7