Add a new parameter to the constrained integer dissector so that it also handles constrained integers with an extension marker.
Update all calls to the constrained integer dissector

Add dissection to the rfc_number type  which is a constrasined integer with an extension marker

Add H245 so that it builds by default in ethereal.
It has been tested extensively by a semi-large number of people with a lot of real and synthetic captures  and seems to work very well.

New protocol added to ethereal


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8032 f5534014-38df-0310-8fa8-9805f1628bb7