Update to the Response Time Statistics taps. SMB/FC/ONC-RPC/DCE-RPC now all use the default tap filter string as the same filter string as is used in the main window instead of using a default NULL filter string.
The idea is that if you have applied a certain filter to your main window, it
is likely that if you want to invoke the response time statistics feature you
probably want to do the response time statistics over the same set of packets, i.e. the ones you see in the main window.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8192 f5534014-38df-0310-8fa8-9805f1628bb7