Due to popular demand:
The heuristic for RPC detection sometimes mistakenly decodes something as RPC when in reality it is payload.

Since the heuristics is based on the PDU suze as specified in the record marker
make this limit settable from the preferences.

Arbitrary limit of 256kb is choosen as default instead of the original limit of 2mb.
(face it, the largest common oncrpc pdus are those for nfs read/write anyway and these seldom go above 32kb.)


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8250 f5534014-38df-0310-8fa8-9805f1628bb7