From Richard Urwin:
	use Export and Import for the buttons in GTK+ 2.x as well;

	get rid of a duplicate fclose;

	other fixes.

Update the description of color filters in the Ethereal man page to
reflect the change, clean up the formatting (use =item), and add the
global and personal color filters files to the FILES section; refer to
them as "color filters" files rather than "colorfilters" files, as the
FILES section gives the "colorfilters" file name so you don't have to
use that as the name.

Clean up white space.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8285 f5534014-38df-0310-8fa8-9805f1628bb7