Update to the conversations list.
Rename it from Endpoint Talkers to : Conversation List

Change command line arguments to both tethereal and ethereal
to be -z conv,<type>
to reflect the new name Conversations

This is the last time the tethereal cmd line arg is changed.  But now it has a
proper intuitive name at least.

io,users was weird
talkers was too close to names used in other tools


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8379 f5534014-38df-0310-8fa8-9805f1628bb7