Add Find Next/Find Previous submenu to conversation lists.
By using Find Next/Previous you will jump to the next/previous matching packet in the ethereal main window.

I could not get CTRL-N / CTRL-B to work and was too lazy to research.
It would be nice if CTRL-N / CTRL-P would invoke the same thing as selecting
/Find Frame/Find Next/EP1 <-> EP2
or
/Find Frame/Find Previous/EP1 <-> EP2
from the menu.
I could not figure out how to get gtk to do this.
The person that adds CTRL-N/CTRL-B here will be a hero.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8635 f5534014-38df-0310-8fa8-9805f1628bb7