Catch exceptions when dissecting a verifier, so we still dissect the stub data even if there's a problem dissecting the verifier.
Show stub data as "Encrypted stub data" if it's encrypted, "Decrypted
stub data" if it was encrypted but we decrypted it, and "Stub data" if
it wasn't encrypted.

Don't attempt to decrypt data unless it was encrypted (i.e., the
authentication level is "Packet privacy".

Get rid of "decrypted_data" member of "packet_info" structure - we don't
need it any more.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@8743 f5534014-38df-0310-8fa8-9805f1628bb7