Performance fix
In the GPROF logs proto_registrar_get_nth() used to take anything between 2.5 and 5.5% of the time.

Replace the GLIB array with a handroleld one for one of the private structures.

the function should now be virtually zero cost
and thus ethereal should be 2.5-5.5% faster on those traces.

anyone that wants to, please rerun GPROF with this fix and see what has changed.


git-svn-id: http://anonsvn.wireshark.org/wireshark/trunk@9058 f5534014-38df-0310-8fa8-9805f1628bb7